Rename iGyne repository and Update Needle Segmentation algorithm
This commit is follow up of 96d3c5a.

It changes both the extension repository and extension name to an
implementation agnostic name.

It implements the last version of the needle segmentation algorithm as well as a new method for CAD model registration.

Debug comments have been disable from printing in the python console.