ENH: Addition of the extension ModelToModelDistance
This extension contains one CLI module of the same name. It computes a point by point distance between two models loaded in Slicer or VTK volumes. It is based on vtkDistancePolyDataFilter. The distance can be signed or unsigned.

More information can be found on the documentation page of the extension:
http://www.slicer.org/slicerWiki/index.php/Documentation/Nightly/Extensions/ModelToModelDistance

Or on the NITRC page of the project:
https://www.nitrc.org/projects/meshmetric3d/

The source code is available on github:
https://github.com/NIRALUser/3DMetricTools