ENH: Adding new extension MeshToLabelMap
MeshToLabelMap computes a label map from a model. Slicer contains a module called "Model to labelmap" module that performs the same operation but that has some limitations due to the asumptions it makes. For more information about the available (and upcoming) modules to perform a model to label map conversion, see discussion [1].

[1] http://massmail.spl.harvard.edu/public-archives/slicer-devel/2014/016831.html