ENH: Addition of ResampleDTIlogEuclidean as an independent extension
ResampleDTIlogEuclidean is a CLI module that performs Diffusion Tensor Image resampling in the log-Euclidean framework (for more information, check [1] and [2]). It is an improvement over ResampleDTIVolume that is included in Slicer which performs the same operation but not in the log-Euclidean framework. ResampleDTIlogEuclidean is currently available in the DTIAtlasBuilder extension. Since Slicer handles extension dependencies since Slicer 4.4, it makes more sense to separate ResampleDTIlogEuclidean and make it a separate extension. This will make it easier for users to find this module and will allow users that only want this specific module and not the whole DTIAtlasBuilder extension to install only what they need.

    [1] Arsigny, V., Fillard, P., Pennec, X. and Ayache, N. (2006), Log-Euclidea
n metrics for fast and simple calculus on diffusion tensors. Magn Reson Med, 56:
 411–421. doi: 10.1002/mrm.20965
    [2] D.C. Alexander, C. Pierpaoli, P.J. Basser, J.C. Gee. Spatial Transformat
ions of Diffusion Tensor Magnetic Resonance Images, IEEE Transactions on medical
 imaging, vol. 20, No. 11, November 2001