ENH: Add AnglePlanesExtension extension
Description:
This Module is used to calculate the angle between two planes by using
the normals. The user gets the choice to use two planes which are
already implemented on Slicer or they can define a plane by using
landmarks (at least 3 landmarks). Plane can also be saved to be reused
for other models.

Contributors:
Julia Lopinto (University of Michigan)