ENH: Add OpenCVExample extension
The OpenCVExample extension provide a template for other extensions
to follow when building and running against the SlicerOpenCV extension.
It provides the BlendVectorVolumes CLI that uses the cv::addWeighted
method to blend together two vector volumes
The extension and module documentation pages on the wiki provide
details on how to use this extension as a template to use opencv
via the SlicerOpenCV extension.