Add files via upload ( + extension Calculator ) (#1662)
* Add files via upload

Hello,  this is a new extention, it consists in a volumetric evaluation of kidney stones and also provides information on surgical duration. Moreover, it provides the estimated time of lithotripsy and consequently the operative time. It provides in a few seconds a 3D view of the stone(s) and axial, coronal and sagittal views with CT-scan DICOMs.
Br

* Delete Slicer-Calculator.s4ext

* Add files via upload

* Update SlicerKidneyStoneCalculator.s4ext