ENH: Add SlicerCIP extension
Description:
Chest Imaging Platform is an extension for quantitative CT imaging
biomarkers for lung diseases. This work is funded by the National Heart,
Lung, And Blood Institute of the National  Institutes of Health under
Award Number R01HL116931. The content is solely the responsibility of
the authors and does not necessarily represent the official views of the
National Institutes of Health.

Contributors:
Applied Chest Imaging Laboratory, Brigham and Women's Hospital