Add the new Similarity Subgroups Extension to the Index
This method not only finds surface deviations of a model, but also shows multiple high accuracy zones (with deviations less than a definable threshold) of a potentially corrupted model. We used it for the validation of 32 models that were used for surgery planning.
The advantage to standard validation methods is, that the surgeon can select a subgroup that fits his region of interest, and can be sure the model is save to use, even when it may be corrupt at other regions.