Archive WASP extension
The extension has not been updated for many years. It was only compatible with Python2.
Probably updating it to Python3 would not have been a lot of work, but since watershed segmentation is already available in SegmentEditorExtraEffects extension, no significant functionality is lost.
The only advantage of WASP that it can precompute multiple watershed segmentations with different settings, so the user can switch between them quickly. This feature could be added to the SegmentEditorExtraEffects extension, too.