Add SlicerNeuro extension
While translating Slicer to different languagues we noticed that a very specialized brain coordinate system transformation module (ACPCTransform) is still in Slicer core. This is an outlier, because all other domain-specific modules are now in extensions. To make translation and maintenance of Slicer core easier it makes sense to move ACPCTransform to an extension.

Since there was no existing extension for generic neuroimaging modules, a new SlicerNeuro extension has been added.

Since the extension only contains one module, it would not be that useful, so some other neuroimaging related extensions are added as dependencies, so the extension can serve as a starting point for users in this field (they need to install only one extension to get all). If more modules will be added to the extension then this may be revisited (extensions that are not actually required can be removed from the extension dependency list).