Do not require "good" extension name for Tier 1
It is often hard to judge if an extension's name is appropriate, when currently the scope of the extension is narrow but developers plan to broaden it in the future.

For example, `SegmentHumanBody` name is extremely broad, but it could be justified for an extension that proves to be universally useful for segmenting anything in the human body (such as TotalSegmentator or MONAIAuto3DSeg).

This commit would allow accepting an extension with a very ambitious name as Tier 1. Since Tier 1 extensions will not be shown by default, an extension with a too ambitious name would not cause disappointments.